Fixed feeds/couchdb so schema load works

+++ config/initializers/couchdb.rb
@@ -1,21 +1,23 @@
-d = Document.new(:name => "temp")
-d.save
-if !d.view_exists("all_data_values")
-  d.create_simple_view("all_data_values", 
-  "function(doc) 
-    {
-      if (doc.data && !doc.is_search_doc)
-      {
-        for(row_key in doc.data)
+if not $rails_rake_task
+    d = document.new(:name => "temp")
+    d.save
+    if !d.view_exists("all_data_values")
+      d.create_simple_view("all_data_values", 
+      "function(doc) 
         {
-          row = doc.data[row_key];
-          for(col_key in row)
+          if (doc.data && !doc.is_search_doc)
           {
-            emit(row[col_key], row);
+            for(row_key in doc.data)
+            {
+              row = doc.data[row_key];
+              for(col_key in row)
+              {
+                emit(row[col_key], row);
+              }
+            }
           }
-        }
-      }
-    }")
-end
+        }")
+    end
 
-d.destroy
+    d.destroy
+end
